Get rid of the hideous 'custom float' hack
We can just do our own verification and show an error message if it's
not a valid float. Main problem is probably going to be getting people
who are already used to having to multiply all these input values by 10
to realize that this is no longer necessary. I included a message to
tell them as much, but who's going to read things...

Also some minor GUI improvements to the other 'custom value' popups.