Hide the tweak_key for all but Tweak Settings
Ever wondered what that weird 'h'/'r' in the custom value popup for
Tweak Clothes was? A symptom of how hacked together this method is :(

It was trying to show the i-th tweak key, which only makes sense for
settings tweaks. So I introduced a new class var for tweaks that's only
set to True for GMST/GLOB tweaks.