Clean removed/renamed columns from settings files
0ffe464d2a45645df26553ae6f082996453146d7 renamed the Activation Status
column to Mod Status, but there was no code added for people who had the
'Activation Status' string stuck in their settings files. Either that,
or we introduced an assumption that all columns in settings files are
valid at some later point.

Would manifest itself as a traceback on startup, slightly different
depending on what part of the settings files had the old column.