FO3, FNV: Rewrite BPTD definition
Was entirely broken. Turns out the logic here is mind-numbingly stupid:
There are two types of 'body parts' that can be listed in BPTD. One is
named, the other is unnamed. A named body part has exactly the same
definition as an unnamed one, except that it has a BPTN subrecord in
front of the rest. That means you can't just use a union or something
like that to handle this. Instead, you need to keep track of whether or
not you hit a BPTN subrecord, then check that once you hit the BPNN
subrecord. If you had a BPTN subrecord at that point, don't create a new
body part, just append the BPNN to the existing one. Otherwise, create a
new body part and skip adding BPTN to it when dumping it later.

At least this was fixed in Skyrim.

Probably makes MelGroups slower because I had to extract a _new_object
method from loadData.

Mopy/bash/game/skyrim/records.py:
Tiny fixup: That subrecord is a localized string